requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
croniter==1.0.6
Django==3.1.6
django-filter==2.4.0
django-model-utils==4.1.1
django-rest-framework==0.1.0
django-rq==2.4.0
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
future==0.18.2
idna==2.10
inflection==0.5.1
mysqlclient==2.0.3
natsort==7.1.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
rq==1.7.0
rq-scheduler==0.10.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.3
